+++ wdl-analysis/src/lib.rs
@@ -10,11 +10,15 @@
 #![warn(rustdoc::broken_intra_doc_links)]
 
 mod analyzer;
+pub(crate) mod diagnostics;
+pub mod eval;
 mod graph;
 mod queue;
 mod rayon;
+mod rules;
 pub mod scope;
 pub mod stdlib;
 pub mod types;
 
 pub use analyzer::*;
+pub use rules::*;