+++ wdl-lint/src/lib.rs
@@ -42,6 +42,14 @@ pub use tags::*;
 pub use visitor::*;
 pub use wdl_ast as ast;
 
+/// The reserved rule identifiers that are used by analysis.
+pub const RESERVED_RULE_IDS: &[&str] = &[
+    "UnusedImport",
+    "UnusedInput",
+    "UnusedDeclaration",
+    "UnusedCall",
+];
+
 /// A trait implemented by lint rules.
 pub trait Rule: Visitor<State = Diagnostics> {
     /// The unique identifier for the lint rule.
@@ -80,8 +88,7 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
         Box::<rules::SnakeCaseRule>::default(),
         Box::<rules::MissingRuntimeRule>::default(),
         Box::<rules::EndingNewlineRule>::default(),
-        Box::<rules::PreambleWhitespaceRule>::default(),
-        Box::<rules::PreambleCommentsRule>::default(),
+        Box::<rules::PreambleFormattingRule>::default(),
         Box::<rules::MatchingParameterMetaRule>::default(),
         Box::<rules::WhitespaceRule>::default(),
         Box::<rules::CommandSectionMixedIndentationRule>::default(),
@@ -112,7 +119,9 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
         Box::<rules::ContainerValue>::default(),
         Box::<rules::MissingRequirementsRule>::default(),
         Box::<rules::UnknownRule>::default(),
-        Box::<rules::MisplacedLintDirective>::default(),
+        Box::<rules::MisplacedLintDirectiveRule>::default(),
+        Box::<rules::VersionFormattingRule>::default(),
+        Box::<rules::PreambleCommentAfterVersionRule>::default(),
         Box::<rules::MalformedLintDirectiveRule>::default(),
     ];
 
@@ -129,6 +138,10 @@ pub fn rules() -> Vec<Box<dyn Rule>> {
 
             if !set.insert(r.id()) {
                 panic!("duplicate rule id `{id}`", id = r.id());
+            }
+
+            if RESERVED_RULE_IDS.contains(&r.id()) {
+                panic!("rule id `{id}` is reserved", id = r.id());
             }
         }
     }